StoppedMockClock now accepts and returns fractional seconds
PHP DateTime objects now always carry microseconds - the
StoppedMockClock should therefore always include them in the times
it accepts and returns to ensure the behaviour is consistent with
the real clock. There is a possibility this will cause some strict
equality checks in unit tests to fail. Not treated as a package
breaking release as it only affects testcase code.