Fix the dockblock typehint for InitialisableSingletonTrait::instance()
It was phpdoced as @return object which meant calls to
MyClass::instance() couldn't properly detect the type of object
in IDE etc.

Leave the actual typehint alone for now, to avoid a BC break on
any classes that use the trait but reimplement that method, but
the docblock can change without any issues.