Fix session & logging bugs with invalid UTF8/control chars in user-agent
Previously, if the user-supplied User-Agent string contained invalid
UTF8 characters this would cause two problems:

* The MysqlSession driver (against MySQL 8 in strict mode) would
  fail to insert the session to the database due to the invalid
  content in the user_agent string. This caused it to throw an exception,
  which would cause session_start to fail with "session ID must be a
  string".
* The request logger would output an empty log line - because JSON-
  encoding the request log entry failed and we were not detecting that.

There was also the theoretical risk that the user input could contain
non-printable control characters - although not directly an issue,
this could cause unexpected behaviour if consuming projects tried to
read the database user_agent or parse the log entries without treating
this as user input.

This fix introduces a new StringSanitiser::ensurePrintableUtf8 method
which will silently re-encode ASCII control characters and invalid
UTF8 sequences as `�` characters to make them safe for use.

That is then used for both the session value and the log value of
the User-Agent header.

The raw User-Agent is still used for hashing the session ID: the hash
function just sees this as a byte sequence so the string encoding is
irrelevant. This ensures that sessions created before this commit
will still have the same hash as ones used after it, so the change is
BC.