Don't escape `/` in JSON-encoded log payloads from StackdriverApplicationLogger
The logs contain fairly regular occurrences of `/` in URLs, file paths, etc.
By default PHP json_encode will encode these as `\/` but this is not required
by the JSON spec.

Use JSON_UNESCAPED_SLASHES to leave these as-is which will slightly reduce the
size of the encoded payloads over the wire. Should not affect any JSON-compliant
parser, so this is not a breaking change.