bug: Fix `ObjectPropertyRipper` to handle `stdClass` objects
Currently, attempting to rip from a `stdClass` will fail with
an error that it is not possible to bind a closure to the
scope of an internal object.

It should not normally be necessary to call any of the
`rip` methods for a `stdClass`, because by definition
all props are public so the result will be the same
as `$vars = (array) $object`.

However, it can cause a problem if e.g. calling the
ObjectPropertyRipper recursively to dump an entire
object graph where the tree may include some stdClass
instances. Our methods are typed to accept any kind of
object, so should work with `stdClass` rather than
requiring the caller to check this externally.