fix: MysqlSession should validate session ID format before checking DB
Credential-stuffing and malware bots will commonly attempt requests
with manufactured session cookies e.g. to search for SQL, path or
URL injection vulnerabilities.

Currently, the MysqlSession driver checks all provided session IDs
against the database to validate them. This is *NOT* a security issue,
since we are using PDO's parameters to properly escape the provided
values.

However, it does cause an exception (and therefore an error report)
because the `GET_LOCK` query uses the user-provided session ID as
part of the lock name. If the lock name contains characters that
are not valid in a lock, MySQL will give a syntax error, which in
turn causes a PDOException.

Instead, skip the database lookup and return false immediately
(as for any other invalid session ID) if the ID does not match
the expected length and character set.